+++ packages/_nova-subscribe-to-posts/lib/export.js
@@ -1,8 +1,10 @@
 import {subscribeItem, unsubscribeItem} from './methods.js';
+import Subscribe from './components/Subscribe.jsx';
 import SubscribeButton from './components/SubscribeButton.jsx';
 
 export {
   subscribeItem,
   unsubscribeItem,
+  Subscribe,
   SubscribeButton
 }
